+++ js/components/discussions/thread-form.vue
@@ -4,7 +4,7 @@
         <label for="comment-new-message">{{ _('Comment') }}</label>
         <textarea v-el:textarea id="comment-new-message" v-model="comment" class="form-control" rows="3" required></textarea>
     </div>
-    <button @click="submit" :disabled="this.sending || !this.comment" class="btn btn-primary btn-block pull-right submit-new-message">
+    <button type="submit" :disabled="this.sending || !this.comment" class="btn btn-primary btn-block pull-right submit-new-message">
         {{ _('Submit your comment') }}
     </button>
 </form>